requirements.txt
numpy
pandas
langchain-groq
langchain-google-genai
langchain-google-vertexai
langchain-neo4j
langchain-graph-retriever
langchain-unstructured
langchain-mcp-adapters
langchain-huggingface
langchain-ollama
langchain_deepseek
mcp
langgraph
neo4j~=5.28.0
dvc
mlflow
google-cloud-aiplatform
browser-use
fastapi
uvicorn
google-genai
deepagents 
tavily-python
uvicorn
a2a-sdk
google-adk
python-dotenv
nest-asyncio
click
google-generativeai
httpx
crewai
langchain-community
groq